+++ vms/client/nx_vms_client_core/src/nx/vms/client/core/resource_views/entity_item_model/entity/composition_entity.cpp
@@ -106,7 +106,7 @@ int CompositionEntity::subEntityCount() const
 
 AbstractEntity* CompositionEntity::subEntityAt(int index)
 {
-    if (index >= m_composition.size() || index < 0)
+    if (index >= (int) m_composition.size() || index < 0)
     {
         NX_ASSERT(false, "Invalid index");
         return nullptr;
